third_party/WebKit/LayoutTests/http/tests/inspector/network/network-disable-cache-preloads-expected.txt

CONSOLE MESSAGE: line 29: PASS - 1 resource loaded
Tests disabling cache from inspector and seeing that preloads are not evicted from memory cache.

Page reloaded.
